Add generate.rs
